Project 2 - MMAlloc/MAlloc utilization
The program should provide a function named mm_alloc.

This is the function prototype:

    void *mm_alloc(int num_bytes_to_allocate)

mm_alloc function should obtain space for the user from a block which you
obtain via mmap.  You should obtain the block on the first invocation
and the block size should be some number of pages.  The number of pages in
the block should be obtained from the environment variable named:
    P2_NUM_PAGES